+++ apps/traffic-sign-tool/src/routeTree.gen.ts
@@ -8,14 +8,15 @@
 // You should NOT make any changes in this file as it will be overwritten.
 // Additionally, you should also exclude this file from your linter and/or formatter to prevent it from being checked or modified.
 
-import { Route as LangRouteImport } from './routes/$lang'
-import { Route as LangCheckSignCombinationsRouteImport } from './routes/$lang.check-sign-combinations'
-import { Route as LangIndexRouteImport } from './routes/$lang.index'
-import { Route as LangSignsRouteImport } from './routes/$lang.signs'
-import { Route as LangTaginfoRouteImport } from './routes/$lang.taginfo'
-import { Route as LangWikiRouteImport } from './routes/$lang.wiki'
 import { Route as rootRouteImport } from './routes/__root'
+import { Route as LangRouteImport } from './routes/$lang'
 import { Route as IndexRouteImport } from './routes/index'
+import { Route as LangIndexRouteImport } from './routes/$lang.index'
+import { Route as LangWikiRouteImport } from './routes/$lang.wiki'
+import { Route as LangTaginfoRouteImport } from './routes/$lang.taginfo'
+import { Route as LangSignsQaRouteImport } from './routes/$lang.signs-qa'
+import { Route as LangSignsRouteImport } from './routes/$lang.signs'
+import { Route as LangCheckSignCombinationsRouteImport } from './routes/$lang.check-sign-combinations'
 
 const LangRoute = LangRouteImport.update({
   id: '/$lang',
@@ -42,22 +43,29 @@ const LangTaginfoRoute = LangTaginfoRouteImport.update({
   path: '/taginfo',
   getParentRoute: () => LangRoute,
 } as any)
+const LangSignsQaRoute = LangSignsQaRouteImport.update({
+  id: '/signs-qa',
+  path: '/signs-qa',
+  getParentRoute: () => LangRoute,
+} as any)
 const LangSignsRoute = LangSignsRouteImport.update({
   id: '/signs',
   path: '/signs',
   getParentRoute: () => LangRoute,
 } as any)
-const LangCheckSignCombinationsRoute = LangCheckSignCombinationsRouteImport.update({
-  id: '/check-sign-combinations',
-  path: '/check-sign-combinations',
-  getParentRoute: () => LangRoute,
-} as any)
+const LangCheckSignCombinationsRoute =
+  LangCheckSignCombinationsRouteImport.update({
+    id: '/check-sign-combinations',
+    path: '/check-sign-combinations',
+    getParentRoute: () => LangRoute,
+  } as any)
 
 export interface FileRoutesByFullPath {
   '/': typeof IndexRoute
   '/$lang': typeof LangRouteWithChildren
   '/$lang/check-sign-combinations': typeof LangCheckSignCombinationsRoute
   '/$lang/signs': typeof LangSignsRoute
+  '/$lang/signs-qa': typeof LangSignsQaRoute
   '/$lang/taginfo': typeof LangTaginfoRoute
   '/$lang/wiki': typeof LangWikiRoute
   '/$lang/': typeof LangIndexRoute
@@ -66,6 +74,7 @@ export interface FileRoutesByTo {
   '/': typeof IndexRoute
   '/$lang/check-sign-combinations': typeof LangCheckSignCombinationsRoute
   '/$lang/signs': typeof LangSignsRoute
+  '/$lang/signs-qa': typeof LangSignsQaRoute
   '/$lang/taginfo': typeof LangTaginfoRoute
   '/$lang/wiki': typeof LangWikiRoute
   '/$lang': typeof LangIndexRoute
@@ -76,6 +85,7 @@ export interface FileRoutesById {
   '/$lang': typeof LangRouteWithChildren
   '/$lang/check-sign-combinations': typeof LangCheckSignCombinationsRoute
   '/$lang/signs': typeof LangSignsRoute
+  '/$lang/signs-qa': typeof LangSignsQaRoute
   '/$lang/taginfo': typeof LangTaginfoRoute
   '/$lang/wiki': typeof LangWikiRoute
   '/$lang/': typeof LangIndexRoute
@@ -87,6 +97,7 @@ export interface FileRouteTypes {
     | '/$lang'
     | '/$lang/check-sign-combinations'
     | '/$lang/signs'
+    | '/$lang/signs-qa'
     | '/$lang/taginfo'
     | '/$lang/wiki'
     | '/$lang/'
@@ -95,6 +106,7 @@ export interface FileRouteTypes {
     | '/'
     | '/$lang/check-sign-combinations'
     | '/$lang/signs'
+    | '/$lang/signs-qa'
     | '/$lang/taginfo'
     | '/$lang/wiki'
     | '/$lang'
@@ -104,6 +116,7 @@ export interface FileRouteTypes {
     | '/$lang'
     | '/$lang/check-sign-combinations'
     | '/$lang/signs'
+    | '/$lang/signs-qa'
     | '/$lang/taginfo'
     | '/$lang/wiki'
     | '/$lang/'
@@ -151,6 +164,13 @@ declare module '@tanstack/react-router' {
       preLoaderRoute: typeof LangTaginfoRouteImport
       parentRoute: typeof LangRoute
     }
+    '/$lang/signs-qa': {
+      id: '/$lang/signs-qa'
+      path: '/signs-qa'
+      fullPath: '/$lang/signs-qa'
+      preLoaderRoute: typeof LangSignsQaRouteImport
+      parentRoute: typeof LangRoute
+    }
     '/$lang/signs': {
       id: '/$lang/signs'
       path: '/signs'
@@ -171,6 +191,7 @@ declare module '@tanstack/react-router' {
 interface LangRouteChildren {
   LangCheckSignCombinationsRoute: typeof LangCheckSignCombinationsRoute
   LangSignsRoute: typeof LangSignsRoute
+  LangSignsQaRoute: typeof LangSignsQaRoute
   LangTaginfoRoute: typeof LangTaginfoRoute
   LangWikiRoute: typeof LangWikiRoute
   LangIndexRoute: typeof LangIndexRoute
@@ -179,6 +200,7 @@ interface LangRouteChildren {
 const LangRouteChildren: LangRouteChildren = {
   LangCheckSignCombinationsRoute: LangCheckSignCombinationsRoute,
   LangSignsRoute: LangSignsRoute,
+  LangSignsQaRoute: LangSignsQaRoute,
   LangTaginfoRoute: LangTaginfoRoute,
   LangWikiRoute: LangWikiRoute,
   LangIndexRoute: LangIndexRoute,